Updating port numbre types from short to int. Updating statistics to multipart in a few more places.

+++ src/main/java/net/floodlightcontroller/core/IOFSwitch.java
@@ -551,7 +551,7 @@ public interface IOFSwitch {
      * Get the portBroadcastCacheHits
      * @return
      */
-    public Map<Short, Long> getPortBroadcastHits();
+    public Map<Integer, Long> getPortBroadcastHits();
 
     /**
      * Send a flow statistics request to the switch. This call returns after
